# frontend/dev/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Immigration CRM API
  version: 1.0.0
  description: Multi-tenant CRM system for immigration agencies with role-based access
    control
paths:
...
  /api/v1/clients/stage-counts/:
    get:
      operationId: v1_clients_stage_counts_retrieve
      description: "\n        Returns the count of clients grouped by their stage.\n
        \       \n        Stage values:\n        - LEAD: Lead\n        - FOLLOW_UP:
        Follow Up\n        - CLIENT: Client\n        - CLOSE: Close\n        \n        The
        counts respect role-based filtering:\n        - Consultant/Branch Admin: Only
        clients in their branch\n        - Region Manager: Clients in all branches
        within their region\n        - Super Admin: All clients in their tenant\n
        \       - Super Super Admin: System-wide access\n        \n        All stages
        are included in the response, even if the count is 0.\n        "
      summary: Get client counts by stage
      tags:
      - clients
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientStageCount'
          description: ''
        '401':
          content:
            application/json:
              schema:
                description: Unauthorized - Invalid or missing authentication
          description: ''
        '403':
          content:
            application/json:
              schema:
                description: Forbidden - Insufficient permissions
          description: ''
...
  /api/v1/clients/{id}/profile-picture/:
    get:
      operationId: v1_clients_profile_picture_retrieve
      description: "\n        Get the profile picture for a client.\n        \n        GET
        /api/v1/clients/{id}/profile-picture/\n        "
      summary: Get client profile picture
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - clients
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
        '404':
          content:
            application/json:
              schema:
                description: Profile picture not found
          description: ''
    delete:
      operationId: v1_clients_profile_picture_destroy
      description: "\n        Get the profile picture for a client.\n        \n        GET
        /api/v1/clients/{id}/profile-picture/\n        "
      summary: Get client profile picture
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - clients
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
        '404':
          content:
            application/json:
              schema:
                description: Profile picture not found
          description: ''
    post:
      operationId: v1_clients_profile_picture_create
      description: "\n        Get the profile picture for a client.\n        \n        GET
        /api/v1/clients/{id}/profile-picture/\n        "
      summary: Get client profile picture
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - clients
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
        '404':
          content:
            application/json:
              schema:
                description: Profile picture not found
          description: ''
  /api/v1/clients/{id}/restore/:
    post:
      operationId: v1_clients_restore_create
      description: Restore a previously soft-deleted client. Only administrators can
        restore clients.
      summary: Restore soft-deleted client
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - clients
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientOutput'
          description: ''
        '400':
          content:
            application/json:
              schema:
                description: Bad Request - Client is not soft-deleted
          description: ''
        '401':
          content:
            application/json:
              schema:
                description: Unauthorized
          description: ''
        '403':
          content:
            application/json:
              schema:
                description: Forbidden - Insufficient permissions
          description: ''
        '404':
          content:
            application/json:
              schema:
                description: Not Found - Client does not exist or not soft-deleted
          description: ''
  /api/v1/clients/{id}/timeline/:
    get:
      operationId: v1_clients_timeline_retrieve
      description: "\n        Get chronological timeline of activities for a client.\n
        \       Returns paginated list of activities ordered by created_at (newest
        first).\n        \n        GET /api/v1/clients/{id}/timeline/\n        "
      summary: Get client timeline
      parameters:
      - in: query
        name: activity_type
        schema:
          type: string
        description: Filter by activity type
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - clients
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
                description: Unspecified response body
          description: ''
...
components:
  schemas:
    BlankEnum:
      enum:
      - ''
...
    ClientOutput:
      type: object
      description: |-
        Serializer for client output (GET requests).

        Returns complete client data including related objects.
      properties:
        id:
          type: integer
          readOnly: true
        first_name:
          type: string
          maxLength: 100
        middle_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        gender:
          oneOf:
          - $ref: '#/components/schemas/ClientOutputGenderEnum'
          - $ref: '#/components/schemas/BlankEnum'
        dob:
          type: string
          format: date
          nullable: true
          title: Date of Birth
        phone_number:
          type: string
          maxLength: 15
        email:
          type: string
          format: email
          maxLength: 254
        referred_by:
          type: string
          description: Referral source
          maxLength: 100
        street:
          type: string
          maxLength: 100
        suburb:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 100
        postcode:
          type: string
          maxLength: 20
        country:
          $ref: '#/components/schemas/CountryEnum'
        visa_category:
          type: integer
          nullable: true
          description: Primary visa category of interest
        visa_category_name:
          type: string
          readOnly: true
          nullable: true
        agent:
          type: integer
          nullable: true
          description: Referring agent (if applicable)
        agent_name:
          type: string
          readOnly: true
        description:
          type: string
        assigned_to:
          type: integer
          nullable: true
          description: Consultant assigned to this client
        assigned_to_name:
          type: string
          readOnly: true
        stage:
          description: |-
            Current stage in client workflow

            * `LEAD` - Lead
            * `FOLLOW_UP` - Follow Up
            * `CLIENT` - Client
            * `CLOSE` - Close
          oneOf:
          - $ref: '#/components/schemas/ClientOutputStageEnum'
          - $ref: '#/components/schemas/BlankEnum'
        active:
          type: boolean
          description: Whether client is actively being serviced
        branch:
          type: integer
          nullable: true
          description: Branch managing this client
        branch_name:
          type: string
          readOnly: true
        created_by:
          type: integer
          readOnly: true
          nullable: true
          description: User who created this record
        created_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when record was created
        updated_by:
          type: integer
          readOnly: true
          nullable: true
          description: User who last updated this record
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when record was last updated
      required:
      - agent_name
      - assigned_to_name
      - branch_name
      - country
      - created_at
      - created_by
      - created_by_name
      - first_name
      - id
      - updated_at
      - updated_by
      - visa_category_name
    ClientOutputGenderEnum:
      enum:
      - MALE
      - FEMALE
      - OTHER
      type: string
      description: |-
        * `MALE` - Male
        * `FEMALE` - Female
        * `OTHER` - Other
    ClientOutputStageEnum:
      enum:
      - LEAD
      - FOLLOW_UP
      - CLIENT
      - CLOSE
      type: string
      description: |-
        * `LEAD` - Lead
        * `FOLLOW_UP` - Follow Up
        * `CLIENT` - Client
        * `CLOSE` - Close
    ClientStageCount:
      type: object
      description: |-
        Serializer for client stage count response.

        Returns counts of clients grouped by stage, respecting role-based filtering.

        Stage values: LE (Lead), FU (Follow Up), CT (Client), CL (Close)
      properties:
        LEAD:
          type: integer
          readOnly: true
          default: 0
          description: Lead count
        FOLLOW_UP:
          type: integer
          readOnly: true
          default: 0
          description: Follow Up count
        CLIENT:
          type: integer
          readOnly: true
          default: 0
          description: Client count
        CLOSE:
          type: integer
          readOnly: true
          default: 0
          description: Close count
        TOTAL:
          type: integer
          readOnly: true
          default: 0
          description: Total client count
      required:
      - CLIENT
      - CLOSE
      - FOLLOW_UP
      - LEAD
      - TOTAL
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT